requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
jinja2
starlette
requests
langchain
langchain-google-genai
jose
typing
python-multipart
langchain_core
pydantic
